MDLE-1723 Make radio buttons more accessible
Before this change the radio buttons to select a course had no label so on a screen reader it would not be obvious what they are for.

After this change the course full name is the label for the radio button (clicking the full name will also now select the course for copying). I also moved the full name to be next to the button, so that the relationship was more logical.