MDLE-2310 The external API has been namespaced
In Moodle 4.2 the external API was moved into a namespace from its old externallib file. This allows it to take advantage of class autoloading and the like.

For details see: https://moodledev.io/docs/4.2/devupdate#external-api

It also appears that in the past the old external text formatting method appears to have allowed an id, rather than a full context object. Now however it only accepts a full context.